+++ NEWS.md
@@ -1,3 +1,8 @@
+# tatoo 1.0.7
+
+* as_Composite_table.data.table converts data.tables to Composite tables, based
+  on separators in the column names (e.g fruit.apple, fruit.banana)
+
 # tatoo 1.0.6
 
 * Removed dplyr dependency